Organizing headers/build for SDK
