+++ CMake/CMakeLibs.cmake
@@ -6,7 +6,6 @@
 
 # osquery-specific helper macros
 macro(SET_OSQUERY_COMPILE TARGET)
-  set_target_properties(${TARGET} PROPERTIES COMPILE_FLAGS ${OS_COMPILE_FLAGS})
   set(OPTIONAL_FLAGS ${ARGN})
   list(LENGTH OPTIONAL_FLAGS NUM_OPTIONAL_FLAGS)
   if(${NUM_OPTIONAL_FLAGS} GREATER 0)
@@ -14,12 +13,12 @@ macro(SET_OSQUERY_COMPILE TARGET)
   endif()
 endmacro(SET_OSQUERY_COMPILE)
 
-macro(ADD_OSQUERY_TEST TEST_NAME SOURCE)
-  if(NOT DEFINED ENV{SKIP_TESTS})
+macro(ADD_OSQUERY_TEST IS_CORE TEST_NAME SOURCE)
+  if(NOT DEFINED ENV{SKIP_TESTS} AND (${IS_CORE} OR NOT OSQUERY_BUILD_SDK_ONLY))
     add_executable(${TEST_NAME} ${SOURCE})
-    TARGET_OSQUERY_LINK_WHOLE(${TEST_NAME} libosquery_basic)
-    set(TEST_FLAGS ${ARGN})
-    if(${TEST_FLAGS})
+    TARGET_OSQUERY_LINK_WHOLE(${TEST_NAME} libosquery)
+    set(TEST_LINK_ADDITIONAL ${ARGN})
+    if(NOT ${IS_CORE})
       target_link_libraries(${TEST_NAME} libosquery_additional)
     endif()
     target_link_libraries(${TEST_NAME} gtest)
@@ -29,13 +28,13 @@ macro(ADD_OSQUERY_TEST TEST_NAME SOURCE)
 endmacro(ADD_OSQUERY_TEST)
 
 # Core/non core link helping macros (tell the build to link ALL).
-macro(ADD_OSQUERY_LINK LINK)
-  ADD_OSQUERY_LINK_INTERNAL("${LINK}" "${ARGN}" OSQUERY_ADDITIONAL_LINKS)
+macro(ADD_OSQUERY_LINK IS_CORE LINK)
+  if(${IS_CORE})
+    ADD_OSQUERY_LINK_INTERNAL("${LINK}" "${ARGN}" OSQUERY_LINKS)
+  elseif(NOT OSQUERY_BUILD_SDK_ONLY)
+    ADD_OSQUERY_LINK_INTERNAL("${LINK}" "${ARGN}" OSQUERY_ADDITIONAL_LINKS)
+  endif()
 endmacro(ADD_OSQUERY_LINK)
-
-macro(ADD_OSQUERY_CORE_LINK LINK)
-  ADD_OSQUERY_LINK_INTERNAL("${LINK}" "${ARGN}" OSQUERY_LINKS)
-endmacro(ADD_OSQUERY_CORE_LINK)
 
 macro(ADD_OSQUERY_LINK_INTERNAL LINK LINK_PATHS LINK_SET)
   if(NOT "${LINK}" MATCHES "(^-.*)")
@@ -55,38 +54,36 @@ macro(ADD_OSQUERY_LINK_INTERNAL LINK LINK_PATHS LINK_SET)
 endmacro(ADD_OSQUERY_LINK_INTERNAL)
 
 # Core/non core lists of target source files.
-macro(ADD_OSQUERY_LIBRARY TARGET)
-  add_library(${TARGET} OBJECT ${ARGN})
-  add_dependencies(${TARGET} libglog)
-  SET_OSQUERY_COMPILE(${TARGET})
-  list(APPEND OSQUERY_ADDITIONAL_SOURCES $<TARGET_OBJECTS:${TARGET}>)
-  set(OSQUERY_ADDITIONAL_SOURCES ${OSQUERY_ADDITIONAL_SOURCES} PARENT_SCOPE)
+macro(ADD_OSQUERY_LIBRARY IS_CORE TARGET)
+  if(${IS_CORE} OR NOT OSQUERY_BUILD_SDK_ONLY)
+    add_library(${TARGET} OBJECT ${ARGN})
+    add_dependencies(${TARGET} libglog)
+    SET_OSQUERY_COMPILE(${TARGET} "-static")
+    if(${IS_CORE})
+      list(APPEND OSQUERY_SOURCES $<TARGET_OBJECTS:${TARGET}>)
+      set(OSQUERY_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
+    else()
+      list(APPEND OSQUERY_ADDITIONAL_SOURCES $<TARGET_OBJECTS:${TARGET}>)
+      set(OSQUERY_ADDITIONAL_SOURCES ${OSQUERY_ADDITIONAL_SOURCES} PARENT_SCOPE)
+    endif()
+  endif()
 endmacro(ADD_OSQUERY_LIBRARY TARGET)
 
-macro(ADD_OSQUERY_CORE_LIBRARY TARGET)
-  add_library(${TARGET} OBJECT ${ARGN})
-  add_dependencies(${TARGET} libglog)
-  SET_OSQUERY_COMPILE(${TARGET})
-  list(APPEND OSQUERY_SOURCES $<TARGET_OBJECTS:${TARGET}>)
-  set(OSQUERY_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
-endmacro(ADD_OSQUERY_CORE_LIBRARY TARGET)
-
 # Core/non core lists of target source files compiled as ObjC++.
-macro(ADD_OSQUERY_OBJCXX_LIBRARY TARGET)
-  add_library(${TARGET} OBJECT ${ARGN})
-  add_dependencies(${TARGET} libglog)
-  SET_OSQUERY_COMPILE(${TARGET} "${OBJCXX_COMPILE_FLAGS}")
-  list(APPEND OSQUERY_ADDITIONAL_SOURCES $<TARGET_OBJECTS:${TARGET}>)
-  set(OSQUERY_ADDITIONAL_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
+macro(ADD_OSQUERY_OBJCXX_LIBRARY IS_CORE TARGET)
+  if(${IS_CORE} OR NOT OSQUERY_BUILD_SDK_ONLY)
+    add_library(${TARGET} OBJECT ${ARGN})
+    add_dependencies(${TARGET} libglog)
+    SET_OSQUERY_COMPILE(${TARGET} "${C_COMPILE_FLAGS} ${OBJCXX_COMPILE_FLAGS} -static")
+    if(${IS_CORE})
+      list(APPEND OSQUERY_SOURCES $<TARGET_OBJECTS:${TARGET}>)
+      set(OSQUERY_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
+    else()
+      list(APPEND OSQUERY_ADDITIONAL_SOURCES $<TARGET_OBJECTS:${TARGET}>)
+      set(OSQUERY_ADDITIONAL_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
+    endif()
+  endif()
 endmacro(ADD_OSQUERY_OBJCXX_LIBRARY TARGET)
-
-macro(ADD_OSQUERY_CORE_OBJCXX_LIBRARY TARGET)
-  add_library(${TARGET} OBJECT ${ARGN})
-  add_dependencies(${TARGET} libglog)
-  SET_OSQUERY_COMPILE(${TARGET} "${OBJCXX_COMPILE_FLAGS}")
-  list(APPEND OSQUERY_SOURCES $<TARGET_OBJECTS:${TARGET}>)
-  set(OSQUERY_SOURCES ${OSQUERY_SOURCES} PARENT_SCOPE)
-endmacro(ADD_OSQUERY_CORE_OBJCXX_LIBRARY TARGET)
 
 # Helper to abstract OS/Compiler whole linking.
 macro(TARGET_OSQUERY_LINK_WHOLE TARGET OSQUERY_LIB)
