+++ CMake/FindGlog.cmake
@@ -1,4 +1,7 @@
 include(FindPackageHandleStandardArgs)
+if(POLICY CMP0054)
+  cmake_policy(SET CMP0054 NEW)
+endif()
 
 set(GLOG_ROOT_DIR "${CMAKE_BINARY_DIR}/third-party/glog")
 set(GLOG_SOURCE_DIR "${CMAKE_SOURCE_DIR}/third-party/glog")
@@ -17,7 +20,7 @@ if(NOT APPLE)
   endif()
 endif()
 
-set(GLOG_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-unnamed-type-template-args")
+set(GLOG_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-unnamed-type-template-args -DHAVE_PREAD")
 
 INCLUDE(ExternalProject)
 ExternalProject_Add(
@@ -32,6 +35,7 @@ ExternalProject_Add(
   INSTALL_COMMAND make install
   LOG_CONFIGURE ON
   LOG_INSTALL ON
+  LOG_BUILD ON
 )
 
 set(GLOG_INCLUDE_DIR "${GLOG_ROOT_DIR}/include")
