+++ CMakeLists.txt
@@ -4,46 +4,51 @@ set(CMAKE_C_COMPILER "/usr/bin/clang")
 set(CMAKE_CXX_COMPILER "/usr/bin/clang++")
 if(APPLE)
   set(APPLE_MIN_ABI "10.9")
-  set(OS_COMPILE_FLAGS "-std=c++11 -stdlib=libc++ -mmacosx-version-min=${APPLE_MIN_ABI}")
-  # Special compile flags for Objective-C++
-  set(OBJCXX_COMPILE_FLAGS "-x objective-c++ -fobjc-arc -Wno-c++11-extensions")
+  set(CXX_COMPILE_FLAGS "${CXX_COMPILE_FLAGS} -std=c++11 -stdlib=libc++")
+  set(CXX_COMPILE_FLAGS "${CXX_COMPILE_FLAGS} -mmacosx-version-min=${APPLE_MIN_ABI}")
   set(OS_WHOLELINK_PRE "-Wl,-force_load")
   set(OS_WHOLELINK_POST "")
+  # Special compile flags for Objective-C++
+  set(OBJCXX_COMPILE_FLAGS "-x objective-c++ -fobjc-arc -Wno-c++11-extensions")
 else()
   if(${CMAKE_SYSTEM_NAME} MATCHES "FreeBSD")
     set(FREEBSD TRUE)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -stdlib=libc++")
+    set(CXX_COMPILE_FLAGS "${CXX_COMPILE_FLAGS} -std=c++11 -stdlib=libc++")
     set(OS_WHOLELINK_PRE "")
     set(OS_WHOLELINK_POST "")
   else()
     set(LINUX TRUE)
-    set(CMAKE_C_FLAGS "-fno-omit-frame-pointer")
-    set(OS_COMPILE_FLAGS "-std=c++11")
+    set(CXX_COMPILE_FLAGS "-std=c++11")
     set(OS_WHOLELINK_PRE "-Wl,-whole-archive")
     set(OS_WHOLELINK_POST "-Wl,-no-whole-archive")
   endif()
 endif()
 
+# make debug (environment variable from Makefile)
 if(DEFINED ENV{DEBUG})
   set(CMAKE_BUILD_TYPE "Debug")
-  set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -g -DDEBUG -O0")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -DDEBUG -O0")
+  set(C_COMPILE_FLAGS "${CMAKE_C_FLAGS} -g -DDEBUG -O0")
 else()
-  set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -O2")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O2")
+  set(C_COMPILE_FLAGS "${CMAKE_C_FLAGS} -O2")
 endif()
 
+# make analyze (environment variable from Makefile)
 if(DEFINED ENV{ANALYZE})
   set(CMAKE_CXX_COMPILER "${CMAKE_SOURCE_DIR}/tools/analysis/clang-analyze.sh")
 endif()
 
+# make sanitize (environment variable from Makefile)
 if(DEFINED ENV{SANITIZE})
-  set(CMAKE_CXX_FLAGS "-g -O0 -fno-omit-frame-pointer")
-  set(OS_COMPILE_FLAGS "${OS_COMPILE_FLAGS} -fsanitize=leak -fsanitize=address")
+  set(CXX_COMPILE_FLAGS "-g -O0 -fno-omit-frame-pointer")
+  set(CXX_COMPILE_FLAGS "${CXX_COMPILE_FLAGS} -fsanitize=leak -fsanitize=address")
   if(LINUX)
-    set(CMAKE_CXX_FLAGS "${OS_COMPILE_FLAGS} -fsanitize=memory")
+    set(CXX_COMPILE_FLAGS "${CXX_COMPILE_FLAGS} -fsanitize=memory")
   endif()
 endif()
+
+# Finished setting compiler/compiler flags.
+set(CMAKE_CXX_FLAGS "${C_COMPILE_FLAGS} ${CXX_COMPILE_FLAGS}"
+  CACHE STRING "compile flags" FORCE)
 
 project(OSQUERY)
 
@@ -59,7 +64,7 @@ string(TOUPPER "${PLATFORM}" PLATFORM)
 list(GET PLATFORM 0 OSQUERY_BUILD_OS)
 list(GET PLATFORM 1 OSQUERY_BUILD_DISTRO)
 
-# Make sure deps were built before compiling
+# Make sure deps were built before compiling (else show warning)
 execute_process(
   COMMAND "${CMAKE_SOURCE_DIR}/tools/provision.sh" check "${CMAKE_BINARY_DIR}"
   WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}"
@@ -67,28 +72,47 @@ execute_process(
   RESULT_VARIABLE OSQUERY_DEPS_CHECK
   OUTPUT_STRIP_TRAILING_WHITESPACE
 )
-
 string(ASCII 27 Esc)
 if(OSQUERY_DEPS_CHECK)
   message(WARNING "${Esc}[31m${OSQUERY_DEPS_MESSAGE}${Esc}[m")
 endif()
 
+# Generate version from git
+execute_process(
+  COMMAND git describe --tags HEAD --always
+  WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}"
+  OUTPUT_VARIABLE OSQUERY_BUILD_VERSION
+  OUTPUT_STRIP_TRAILING_WHITESPACE
+)
+
+# make extensions (tests building SDK-based extensions)
+if(DEFINED ENV{SDK})
+  set(OSQUERY_BUILD_SDK_ONLY TRUE)
+  add_definitions(
+    -DOSQUERY_BUILD_SDK
+    -DOSQUERY_BUILD_SDK_VERSION=${OSQUERY_BUILD_VERSION}
+  )
+endif()
+
+# Set various platform/platform-version/build version/etc defines.
+add_definitions(
+  -DOSQUERY_BUILD_VERSION=${OSQUERY_BUILD_VERSION}
+  -D${OSQUERY_BUILD_OS}
+  -D${OSQUERY_BUILD_OS}_${OSQUERY_BUILD_DISTRO}
+)
+
 if(APPLE)
   message("-- Building for OS X")
 elseif(OSQUERY_BUILD_OS STREQUAL "UBUNTU")
-  set(LINUX TRUE)
   set(UBUNTU TRUE)
   message("-- Building for Ubuntu")
 elseif(OSQUERY_BUILD_OS STREQUAL "CENTOS")
-  set(LINUX TRUE)
   set(CENTOS TRUE)
   message("-- Building for CentOS")
 elseif(FREEBSD)
   message("-- Building for FreeBSD")
 endif()
 
-set(USER_COMPILE_FLAGS "-static")
-set(OS_COMPILE_FLAGS "${OS_COMPILE_FLAGS} ${USER_COMPILE_FLAGS}")
 set(CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/CMake" "${CMAKE_MODULE_PATH}")
 set(CMAKE_SKIP_INSTALL_ALL_DEPENDENCY TRUE)
 
@@ -119,6 +143,7 @@ link_directories("/usr/local/lib")
 enable_testing()
 
 include(CMakeLibs)
+include(Packages)
 add_subdirectory(osquery)
 
 # make docs
@@ -140,6 +165,7 @@ add_custom_target(
   COMMENT "Formatting all osquery code with clang-format" VERBATIM
 )
 
+# make format
 add_custom_target(
   format
   python "${CMAKE_SOURCE_DIR}/tools/formatting/git-clang-format.py"
