+++ Makefile
@@ -31,6 +31,14 @@ sanitize: .setup
 	cd build/$(BUILD_DIR) && SANITIZE=True cmake ../../ && \
 	  $(MAKE) --no-print-directory $(MAKEFLAGS)
 
+sdk: .setup
+	cd build/$(BUILD_DIR) && SDK=True cmake ../../ && \
+	  $(MAKE) --no-print-directory $(MAKEFLAGS)
+
+test_sdk: .setup
+	cd build/$(BUILD_DIR) && SDK=True cmake ../../ && \
+	  $(MAKE) test --no-print-directory $(MAKEFLAGS)	
+
 deps: .setup
 	./tools/provision.sh build build/$(BUILD_DIR)
