+++ include/osquery/database/query.h
@@ -14,12 +14,15 @@
 #include <memory>
 #include <string>
 
-#include <gtest/gtest_prod.h>
-
 #include <osquery/database/db_handle.h>
 #include <osquery/database/results.h>
 #include <osquery/scheduler.h>
 #include <osquery/status.h>
+
+#ifndef FRIEND_TEST
+#define FRIEND_TEST(test_case_name, test_name) \
+  friend class test_case_name##_##test_name##_Test
+#endif
 
 namespace osquery {
