+++ osquery/devtools/CMakeLists.txt
@@ -1,6 +1,6 @@
-ADD_OSQUERY_CORE_LIBRARY(osquery_devtools
+ADD_OSQUERY_LIBRARY(FALSE osquery_devtools
   shell.cpp
   printer.cpp
 )
 
-ADD_OSQUERY_TEST(printer_tests printer_tests.cpp)
+ADD_OSQUERY_TEST(FALSE printer_tests printer_tests.cpp)
