--- osquery/examples/CMakeLists.txt
@@ -1,1 +1,0 @@
-ADD_OSQUERY_TEST(example_test example_test.cpp)
