+++ osquery/scheduler/CMakeLists.txt
@@ -1,5 +1,5 @@
-ADD_OSQUERY_CORE_LIBRARY(osquery_scheduler
+ADD_OSQUERY_LIBRARY(TRUE osquery_scheduler
   scheduler.cpp
 )
 
-ADD_OSQUERY_TEST(scheduler_tests scheduler_tests.cpp)
+ADD_OSQUERY_TEST(TRUE scheduler_tests scheduler_tests.cpp)
